db can be queried and habits added to db
