+++ server/db.js
@@ -1,11 +1,9 @@
-var express = require('express');
-var app = express();
 var pg = require('pg');
 module.exports = function(app){
 //========================================================//
 //   Database Routes                                      //
 //========================================================//
-var connectionString = 'postgres://localhost:5432/postgres';
+var connectionString = process.env.DATABASE_URL || 'postgres://localhost:5432/kmerino';
 
 app.get('/api/profile', function(req, res){
    pg.connect(connectionString, function(err, client, done){
@@ -19,34 +17,51 @@ app.get('/api/profile', function(req, res){
     });
     query.on('end', function(result) {
       client.end();
-      // return res.json(rows);
-      return 'cats';
+      return res.json(rows);
+      
+    });
+  });
+});
+// SHOWS EXISTING USER HABITS
+app.get('/api/habits', function(req, res){
+   pg.connect(connectionString, function(err, client, done){
+    var query = client.query('SELECT user_id, habit from habits');
+    var rows = []; // Array to hold values returned from database
+    if (err) {
+      return console.error('error running query', err);
+    }
+    query.on('row', function(row) {
+      rows.push(row);
+    });
+    query.on('end', function(result) {
+      client.end();
+      return res.json(rows);
+      
+    });
+  });
+});
+// USER CREATES A NEW HABIT
+app.post('/api/habits', function(req, res){
+  pg.connect(connectionString, function(err, client, done){
+    var query = client.query("INSERT INTO habits (user_id, habit) VALUES (1, 'fun');");
+    var rows = []; // Array to hold values returned from database
+    if (err) {
+      return console.error('error running query', err);
+    }
+    query.on('row', function(row) {
+      rows.push(row);
+    });
+    query.on('end', function(result) {
+      client.end();
+      return res.json(rows);
     });
   });
 });
 
-app.post('/api/addHabit', function(req, res){
-  pg.connect(connectionString, function(err, client, done){
-    var query = client.query('INSERT INTO habits (user_id, habit) VALUES (1, ‘biking every day’);');
-    // var rows = []; // Array to hold values returned from database
-    if (err) {
-      return console.error('error running query', err);
-    }
-    // query.on('row', function(row) {
-    //   rows.push(row);
-    // });
-    // query.on('end', function(result) {
-    //   client.end();
-    //   return res.json(rows);
-    // });
-  });
-});
-
-
+// USER UPDATES HABITS
 // app.post('/api/updateHabit', function(req, res){
 //   pg.connect(connectionString, function(err, client, done){
 
 
 // });
 }
-// module.exports = db;
